ffmpeg works

+++ camera/src/outputs/ffmpeg_output.py
@@ -1,5 +1,10 @@
+import os
 import subprocess
 import platform
+
+class FFMPEGNotFoundException(Exception):
+    """Raised when FFMPEG can not start or is not found"""
+    pass
 
 class FFMPEGOutput:
     """
@@ -13,10 +18,11 @@ class FFMPEGOutput:
 
         # h264_v4l2m2m is a codec for Arm/Pis which can use hardware acceleration
         output_codec = "h264_v4l2m2m" if is_pi else "mjpeg"
+        hwaccel = self.get_hwaccel()
 
         command = ["ffmpeg",
                    "-re",                   # Real time
-                   "-hwaccel", "drm",       # Use hwaccel if you don't want to suffer
+                   "-hwaccel", hwaccel,     # Use hwaccel if you don't want to suffer
                    "-f", "rawvideo",        # Raw video format (No container like with mjpeg or h264)
                    "-vcodec", "rawvideo",   # Raw video codec needed for rawvideo format
                    "-pix_fmt", "bgr24",     # Input format is bgr24 (blue green red with 8 bit channels)
@@ -31,9 +37,17 @@ class FFMPEGOutput:
         
         self.proc = subprocess.Popen(command, stdin=subprocess.PIPE)
 
+    def get_hwaccel(self):
+        os_name = os.name
+        if os_name == "Windows":
+            return "d3d11va"
+        else:
+            return "drm"
+
     def stop(self):
-        self.proc.stdin.close()
-        self.proc.terminate()
+        if hasattr(self, "proc"):
+            self.proc.stdin.close()
+            self.proc.terminate()
 
     def __del__(self):
         self.stop()
